[flexiglass] Set up initial doze state early
The attached bug occurs because the initial doze amount isn't being set
as soon as the Lockscreen scene is entered.

The fix is simple: set the initial doze amount when changing to the
Lockscreen scene, based on the KeyguardState.

Fix: 379848896
Test: manually verified by pressing the power button to enter AOD while
showing Bouncer but also when being unlocked and showing either Gone or
Shade
Flag: com.android.systemui.scene_container

Change-Id: I389ff51267ea84351b872455ffd5f0f928e3420b

Former-commit-id: 46759e76ca88e5df698b2da6321f0e8dc9091cfb