Hide pattern by default when a non-touch pointing device is available
When drawing a pattern on the lockscreen using the touch screen or a
mouse the pattern is visible by default for some time.
This change modifies the default value on devices that have a pointing
device, like a mouse or touchpad, attached or connected to their device
to hide the drawn pattern by default. Leaving the default for
touch-only input unchanged.

Flag: com.android.internal.widget.flags.hide_last_char_with_physical_input
Bug: 339270220
Test: atest LockPatternUtilsTest

Change-Id: I2142abe2f469bb65a56be68f07990897083b43b4

Former-commit-id: ebe378d25f8fd68bdcaeaefe97a629afc1d4d503