Handle activity as incompatible if whole stack is transparent
Currently we only handle single top transparent activities as
incompatible for desktop windowing. The idea behind this was that if a
transparent activity is on top of another activity, it can be contained
within the window of the activity below. However, if the below activity
is also transparent, when the top activity is contained within it there
will be no background producing a weird floating window.

Instead we should treat an activity as incompatible and force it to
fullscreen if all the activities in the task's stack are transparent.

Flag: com.android.window.flags.enable_desktop_windowing_modals_policy
Test: atest WMShellUnitTests:AppCompatUtilsTest
Test: atest WMShellUnitTests:DesktopTasksControllerTest
Test: atest WMShellUnitTests:DesktopModeWindowDecorViewModelTests
Test: atest WMShellUnitTests:SystemModalsTransitionHandlerTest
Fixes: 379735058
Fixes: 369796795
Change-Id: Ie97286a7553a7318d36e250c82a78ae02b7b78b6

Former-commit-id: 93e246e7cf799dda7fe67d320460a91e3fc559d7