LocationFudger: Bug fix in test.
This test should have the flag enabled, as the method name suggests.
The test was passing despite the wrong value of the flag because the tests checks that `getCoarseningLevel` wasn't called, which is trivially true if the flag is off.
The tests still passes when the flag is on.

This change has no impact on production code, only on tests.

Tests:
- atest FrameworksMockingServicesTests:LocationFudgerTest

Test: manual atest on Pixel 7 pro (see above)

Bug: 382014485

Flag: TEST_ONLY

Change-Id: Ia040d30b0e01aff5b60d5ba54b83d51bf1ddeebb