Allow a frame to receive insets when a side is partially covered
Previously, a frame can only receive insets only when a side is fully
covered by an insets source.

This CL relaxes the restriction. This CL allows a frame to receive
insets when a side is partially covered by the insets source if the side
is the same as the side hint of the source.

Fix: 376941605
Flag: EXEMPT bugfix
Test: atest InsetsSourceTest
Change-Id: Ife05b5a8d2e0eb9e3c891f774de6457d4c398cd4

Former-commit-id: 9e88193231a76f7c5e8b4631a8587737291e447a