Location Fudger: Gate behind correct flags
Fixes the current warning: `[FlaggedApi] Method getCoarseningLevel() is a flagged API and should be inside an if (Flags.populationDensityProvider()) check`

In practice, calling the method while the flag is disabled should never happen, as "Flags.populationDensityProvider()" should imply that the cache is null (see the code in the Manager). Yet, it seems bad practice to call an API gated behind a flag without checking this flag, so in essence the warning is correct. Thus this commit should remove a potential footgun if the code in the manager changes.

Tests:
- atest FrameworksMockingServicesTests:LocationFudgerTest
- atest FrameworksMockingServicesTests:LocationFudgerCacheTest

Test: manual atest on Pixel 7 pro (see above)
Flag: android.location.flags.density_based_coarse_locations
Bug: 376198890
Change-Id: I8bfd9c6e4f4921d7e6389f2924b6b3566a8ae07f

Former-commit-id: 9d9a4d20ec71816f0c6b75aa539303cc8f26ed15