Set a maximum expanded view width for large screens
We previously set a constant width for the expanded view in large
screen devices, but some display configuration may not be able to
fit that width. This change checks that the expanded view can fit
in the screen, and otherwise sets it to fill the screen with equal
padding on both sides.

Flag: EXEMPT bugfix
Fixes: 382256445
Test: atest BubblePositionerTest
Test: manual
       - adb shell wm size 1340x800
       - adb shell wm density 213
       - set device orientation to portrait
       - send floating bubble and expand it
       - observe expanded view fits in screen
Change-Id: Ifbf203cf03fcb1a9f511f8de96a2f37a26d42c6b