Add backup metrics for settings that use the extractRelevantValues() method.
Adds metrics logging for settings that are queried from the db. That is: system, global, secure and device_specific_config. 

This is part of an effort to add metrics support to SettingsBackupAgent. For more details, see go/settings-backup-agent-metrics-design

Bug: 379861078 
Change-Id: If8c881da91902df4d630bf8eba854556bbdf7d37
Flag: com.android.server.backup.enable_metrics_settings_backup_agents
Tested: Manually by verifying that the metrics are logged in GMS Core.

Former-commit-id: 53d643c9427b8a00a41d20d1e24d090d6944051d