Refactor "rearrange" tests to use verifyDisplay
This makes tests in this file more consistent and more thorough. We now
verify that sizes and IDs are not changed by rearrange unintentionally.

The normalize() method now also sorts the children in a predictable
order, and resets unused fields on the root node in order to ease
testability.

Flag: com.android.settings.flags.display_topology_pane_in_display_list
Test: atest DisplayTopologyTest.kt
Bug: b/352650922
Change-Id: I414a52961b1c218939a701af69b0b66020277fd4

Former-commit-id: 9edfba0bde3fcfde65cfe7f71a4b4e91d1dd8774