Minimize oldest task when clicking New Window at window limit.
Wires the multi instance "New Window" button logic into existing task
minimization logic, causing a new task launch from this button to
minimize if the desktop is at the window limit.

Also removes the check against minimizeChange in
DesktopMixedTransitionHandler as we now have a case where a null
launchChange could be paired with a minimizingTask (new window would
have resulted in a minimize but was aborted).

Bug: 378920066
Bug: 380198576
Test: Manual
Flag: EXEMPT bugfix
Change-Id: Icef8e92dbbbd82529e633211d1304f4334353566

Former-commit-id: 37bb11081c88e38a4a0aead84b101be4becea702