[kairos] introduce Incremental type
This type is a specialization of State<Map<K, V>> that models
"incremental" updates to the Map, via a new `MapPatch<K, V>` alias.

There are situations where it is useful to react to only what has
changed between two states; normally this is accomplished by diffing the
two states, but Incremental allows you to avoid performing the diff and
instead operating on the already-known "patch".

Flag: EXEMPT unused
Test: atest kairos-tests
Change-Id: I2829ffc149fb2e1888a11003908068e0ccc8fb35

Former-commit-id: eb057d3eb942520b2e6b0fb5c5d910cd9765a37b