Implement resizing of ongoing content in responsive grid
In a responsive grid, we don't actually know how many rows/columns we
will have until we render the grid. Therefore in order to do dynamic
sizing of ongoing items, it needs to be done after the grid layout. We
therefore move this logic to a function which we can call from the ui
code.

Flag: com.android.systemui.communal_responsive_grid
Test: atest ResizeUtilsTest
Bug: 378171351
Change-Id: I6c1d345d6701b8b1cbb541902e1f1716d5c02717