Use UserTracker to get the correct clipboard manager
Currently we only start the clipboard manager once, so it doesn't work
after switching to a secondary user. This change listens for user
switches and updates the clipboard and keyguard managers to be
associated with the current user.

Note that there is still an issue causing the clipboard overlay window
to not display for secondary users. This is a preliminary change that
just fixes the issue of listening to the correct user.

Bug: 217922018
Test: manual via adding a secondary user, verified via logs
Flag: com.android.systemui.clipboard_overlay_multiuser
Change-Id: Ia9aa6f1a79d8f964c0b3831f901b6c66edeea3d0

Former-commit-id: 6860d18582895f16ee46ed205668045d78778bc2