Refine API of getting window visible display frame
The API of getWindowVisibleDisplayFrame() is using an estimated way to
calculate the frame when the view is not attached to a window. However,
the current estimation uses the navigation bar and display cutout as the
insets covering the view content, which is not consistent with the case
where the view is attached to a window.

Make the necessary change to let it consistent with the attached case.

Bug: 373497540
Test: PopupWindowTest
Flag: EXEMPT BUGFIX
Change-Id: I82df6b2da21bf186a296ce6ae8fffd4ffebb9f6f

Former-commit-id: 4d379cbf7746fc8a14e59f1635c5850ba0e0cd3f