[flexiglass] Update NSSL stack height on child height changes
NSSL has two concepts of stack heights: one for the available space,
and one for the required (intrinsic) stack height.
When a Notification changes its height (e.g. it swaps public to private layout bc the device gets unlocked), we are updating the intrinsic stack height, which is the correct thing to do.
However there are scenarios, when we need to update the available space as well, because the NSSL lays out its  content like "wrap height" instead of "match parent". This update was missing.

Fixes: 382626688
Test: atest NotificationStackScrollLayoutTest
Test: unlock the device with face auth (without bypass) -> notifications update their content correctly
Flag: com.android.systemui.scene_container
Change-Id: I0094a9b34c0ab025df63203a3a38444dba27605c