ExifInterface: do not spam stacktrace on invalid format
If we scan a directory with "unsupported" files, we get spam
stacktraces.

This is Debug at best.

Also simplify the error so it doesn't drift out of date with reality.
ExifInterface now supports reading more than just JPEG and RAW files,
see class-level javadoc for the latest list of supported formats.

Change-Id: I18e0b10ad735c9ce5696a2ad79f7551203c39d92

Former-commit-id: 4eeb4fe9bd1caab47975d4e8583e492181e2d439