Update visible requested of wallpaper with target
Now if the wallpaper target's visibility is changed, it will request
to update wallpaper target. And uses the visible-requested of the
current target to update wallpaper's visible-requested.

This removes invisible-request in commitVisibleWallpapers because
it may make wallpaper invisible unexpectedly with the condition:
Wallpaper target is visible but it is not in mParticipants.

Bug: 355596979
Bug: 354459548
Flag: com.android.window.flags.ensure_wallpaper_in_transitions
Test: WallpaperControllerTests#testWallpaperTokenVisibilityWithTarget

Change-Id: I516efc9c4904585d9d7873e568104d4c2b13bfcf

Former-commit-id: 275024ae65a00437bcc755eddebf40ce25b6bc6f