Lazily inflate the maximize button's progress bar
The progress bar behind the header's maximize button is only needed when
the user hovers over the button, yet it takes 1ms to inflate it in the
shell.main thread in the middle of transition dispatching, which causes
latency.
This change replaces it with a ViewStub that waits to inflate the real
progress bar until it is actually needed.

Bug: 360452034
Flag: com.android.window.flags.enable_desktop_windowing_mode
Test: verify in perfetto that ProgressBar inflation is down to ~35us
Change-Id: I7137b5886a4f012ac80215082a4452ef435ec37b

Former-commit-id: 1efe85914b07decb40c04b3a124d43d996ba41d9