Allow the system user to access public displays
On HSUM devices some activity launches as the system user on public
secondary displays fail (VirtualDeviceHomeTest) because that user is
not considered "visible" by UserVisibilityMediator.

This is already taken into account in DisplayContent#hasAccess but
not in ActivityTaskSupervisor#isCallerAllowedToLaunchOnDisplay.
A similar check is done in Display#hasAccess and both of the existing
methods allow the access for the system user.

This CL is simply making the logic consistent and valdiating the
HSUM support for these activity launches.

Fix: 376008387
Test: atest VirtualDeviceHomeTest
Flag: EXEMPT trvivial bugfix
Change-Id: I708e33c0917a3a589528de2cfaf7edb2aed8c547

Former-commit-id: 97d062ea06d87b8437eb0c13908b5dcb928e710d