Set UMO settings icon to dynamic color on communal
This changes sets the tint of the settings gear icon of UMO to a dynamic
color when UMO is on the communal / glanceable hub, and sets it back to
the previous default color whenever UMO leave communal. This is for
accessibility reason so that the icon has sufficient contrast with the
background of communal, which is a materical dynamic color.

Fix: 378797118
Test: manually tested that the tint of the settings button is set; see
      bug for a video
Flag: EXEMPT simple UI change
Change-Id: Ie27f0026fc35b4438a6e022624d60f52d7ae2b35

Former-commit-id: 4c27a8791be5a3350e5799e146f5b684b0067ede