Allow debug InputManagerGlobal logs to be dynamic
This will allow us to enable these via "DebugInputRule" in tests.

Prior to this CL, the logs were static. That is, whenever
InputManagerGlobal was created, the value was read and stored. The test
process is the same throughout the test. As a result, if an earlier test
resulted in creation of InputManagerGlobal, subsequent tests would reuse
the same value, thus leading to no logs.

Bug: 380535703
Flag: EXEMPT needed for debugging
Test: logcatcolor  | grep -i InputManager
Change-Id: I642b3b5786c73020d08ccbb8817322322a0df911

Former-commit-id: 2c5fb35833c2b32c4539769b9d04267ff76a26e4