AudioDeviceInventory: improve BT device switch time
Add an argument to AudioSystem.setDeviceConnectionState()
to indicate that a device disconnection is part of a disconnect/connect
sequence corresponding to a switch betwen two devices of the same type
(same profile for Bluetooth devices).
This is currently only used for Bluetooth devices.
This tells the audio policy manager that some of the rerouting operations
that normally happen when a device is disconnected can be skipped.

Flag: com.android.media.audio.optimize_bt_device_switch
Bug: 373867402
Test: Bluetooth audio use cases regression
Change-Id: I7848fd541bf4baa529907426034c0dc10ced3773

Former-commit-id: eaa7253d7079689d63060039a9442656674277e6