Implement system session lifecycle callbacks
This CL adds the concept of a global session (note that a default
session exists already), which is the session that's applicable to all
apps unless they have an overriding session (which is a session that
overrides the global session and affects one app specifically).

This means a new capability of the system provider to have one routing
session by default, and then apply specific routing sessions to specific
apps.

Bug: b/362507305
Test: atest CtsMediaBetterTogetherTestCases:android.media.bettertogether.cts.SystemMediaRoutingTest --rerun-until-failure
Flag: com.android.media.flags.enable_mirroring_in_media_router_2
Change-Id: I6c25f75feb200c86f67e569b07cc6cbef53ac8a3

Former-commit-id: 73c6cb3467946dd8470f97adf11375640aed0b6e