Use elapsedRealtimeClock as time source for RateEstimator in NM
InstantSource.system is basically currentTimeMillis(), which means it's not suitable for timing (can jump forward or backwards). For rate estimation, we don't care about correspondence with real-world time, so elapsedRealtime() is what we want.

Also moved the creation of the InstantSource to NotificationManager; SystemServiceRegistry shouldn't need to be concerned about this, nor know the details of the specific time source that is best.

Bug: 381875150
Test: manual
Flag: android.app.nm_binder_perf_throttle_notify
Change-Id: Ic336b8d26777a875476c75950875657265b2fa16

Former-commit-id: 9a6e3f4b7b81d910aceb2efdeb7ad2f7c956c505