Show/hide the animation on PIN entry when changing connected devices
Keyguard would previously decide if it should show the animations on a
button when creating the view. This can lead to unexpected behavior
when attaching or detaching a physical keyboard from the device.

Now Keyguard listens to changes on the connected devices and changes
the animation showing/hiding behavior when necessary.

Flag: com.android.internal.widget.flags.hide_last_char_with_physical_input
Test: atest KeyguardPinBasedInputViewControllerTest
Bug: 384694667

Change-Id: Iab3feda1cc692ae7fbe46623665cb5f8329eaefd

Former-commit-id: cf59cf808df309837f399d60f1808326115f0356