Playing face auth MSDL haptics when lockscreen does not dismiss.
If the user setting to dismiss the keyguard is turned off, we should
still play haptics when face authentication successfully unlocks the
device. This can be done via a combination of the power state button
state upon device entry, and the face authentication result and the
state of the bypass setting.

Test: DeviceEntryHapticsInteractorTest
Flag: com.android.systemui.msdl_feedback
Bug: 361165906

Change-Id: I99b6a3c3d36373e2a6a2f9b2bee414beb6228793

Former-commit-id: 7204034dfa2aa9bb95dcd6f684737d94743a5b91