Introduce latency tracking for shade display change
This tracks the latency of moving the shade between displays, from the
trigger to the end of the first doFrame with the correct configuration.

Note that this is is not perfect: the ideal solution would be tracking
the time until the buffer having the new configuration is presented to
the display, but this would require an extensive rework of how
LatencyTracker works (as we receive the present timestamp, which will be
in the past), so it will be for a follow up cl.

Bug: 362719719
Bug: 378688537
Test: ShadeDisplayChangeLatencyTrackerTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I09db08a9c602338c6720bb321806069d8c62e795

Former-commit-id: 360c7f7a1cd626eff1f828b81fa1f85834687230