Revert^2 "Move Settings backup and restore metric logging from SettingsProvider to SettingsHelper"
This reverts commit 5f82a9bac272e54aff2cbd66db3aeeac3ffdb4de [formerly 5f9760ec36940df3267173717d103ba0dbd80ff0].

Reason for revert: Roll forward with fix. The previous change caused test failures because I didn't update the arguments of a couple of calls of SettingsHelper.restoreValue(...) in a test. Now that the keys are defined in an outer class and the logger is set independently, there's no need to modify any arguments of that method.

Change-Id: I4b6af39d011393c8fb2312024369fc0cf418a921
Bug: 379861078
Flag: com.android.server.backup.enable_metrics_settings_backup_agents
Tested: SettingsHelperTest
Tested: SettingsBackupAgentTest

Former-commit-id: f43ab3285cadf09aa02a4e472ddaf2101ae568e2