Disable calls to setBluetoothA2dpOn()
This change is intended to assertively define the role of the media
routing framework in setBTA2dpOn() bugs. In the future, we might
consider merging making this the real behavior but, until then, we can
also use this change to experiment with real usage.

Bug: b/372908448
Flag: com.android.media.flags.disable_set_bluetooth_ad2p_on_calls
Test: atest CtsMediaBetterTogetherTestCases
Test: Manually verified that regular media usecases using 1p and 3p
Test: popular apps work as expected.
Test: Also verified that MediaRouter APIs (legacy and AndroidX) work
TesT: as expected. With the exception of not calling setBTA2dpOn().
Change-Id: I286d61c585c1e6cb20531c659a09f377f72f95a6

Former-commit-id: fa9d4bafd8ba36b8ac95033961d957350f917351