Add spacer if the last widget is larger than the dragging widget
Previously, dragging a smaller widget beyond the last large widget in
the grid was not possible. This occurred because the grid could not
overscroll to make the target position visible when no additiona space
is available.

This change adds a spacer after the last widget if it is larger than the
dragging widget. This spacer allows overscrolling, enabling the dragging
widget to be positioned beyond the large widget.

Test: manual; see videos in bug
Bug: 383384880
Flag: EXEMPT bug fix
Change-Id: Ic513ded3cf804acedcf9355a0385c13eaf96c5ab

Former-commit-id: 245501c54a1fc26a94f34da0c7e85b801f62ec1b