Setup ANGLE when it's the system driver
If ANGLE is being used as the OpenGL ES driver, either because the
various flags/settings force the selection or it's the system driver
(ro.hardware.egl=angle), it always needs to be setup. This allows
for configuring ANGLE feature overrides via EGL attributes, using ANGLE
drivers within an installed APK, etc.

Bug: 372694741
Test: Boot the device with ANGLE as the system driver.
Change-Id: I3e5f6ec621569eb76605f55e9236b58a4ebcee08

Former-commit-id: bb0884206ddc0d2858b0a64df595b80410990813