[SB][Notif] Only hide text in chip that was actually tapped.
Previously, if any notification chip was tapped, we'd hide the text in
all chips. Instead, we should just hide the text in the chip that was
actually tapped.

Most of this CL is adding a new flow to `HeadsUpNotificationInteractor`
that emits not just the current `PinnedStatus` but also the key for the
pinned row.

Fixes: 386784519
Bug: 364653005
Flag: com.android.systemui.status_bar_notification_chips

Test: atest HeadsUpNotificationInteractorTest NotifChipsViewModelTest
Test: Have 2 notification chips, tap on one -> verify only that chip
hides text. Tap on other -> verify other chip hides text, previous chip
shows text

Change-Id: I7332e40565ed5b6892645168626a78f31461e243

Former-commit-id: 33c2d33358a41ceb0976eee2bcd6ceae2f6f480a