Explicitly keep default constructor in rules without members
R8 implicitly keeps the default constructor when no member rules are specified. Future versions of R8 will require apps to explicitly keep the default constructors that are required for the build to work.

This conservatively changes all keep rules that currently keep the default constructor to explicitly keep the default constructor.

Future work will attempt to tighten up the rules that do not require the default constructor to be kept.

Bug: 373579455
Flag: NONE no behavior change
Test: existing
Change-Id: I8e2734d8ee1b14cdf8ed6e3a1b6ebb8432500176

Former-commit-id: 6dcde63a1ed81bafb04309d701501e00b31d0d4e