Keyguard Transitions - Update check for canceling
Instead of relying on an emitted transition, check the
animator directly to see if a cancel will be issued, which
is the actual source of truth for automated transitions.

Also, fix the transition runner. The handler needs to be
updated once per test run in setup, not for each call to
startTransition, which was causing race conditions.

Also, add a bit more logging for the test where this issue
was detected.

Test: atest SystemUITests:KeyguardTransitionRepositoryTest
Bug: 383318506
Flag: EXEMPT bugfix
Change-Id: I265cb41b1c4494827c91cb98287ab2b7e2102b4a

Former-commit-id: 580a65482fd37162460faedc9b99807fb49af33c