Move open-animation impl from taskview to controller [3/N]
Continue to move high-level control out of taskview impl.

In this case, some animation logic was in taskview internals.
This is probably the most "customized" bit, so broke out the
animation part from the "setting state" and "notifying
listeners" parts.

Again, this was just mechanical code movage, but doing it in
its own CL since it isn't verbatim.

Bug: 384976265
Test: refactor only, existing tests
Flag: EXEMPT mechanical refactor
Change-Id: I3532db1e858519be4ab223605104ab3389c2526f

Former-commit-id: f1842a3aed30f48a3e3843853971e2c1c147461c