Collapse the shade before moving the window to another display
This makes sure that the shade is collapsed before the shade is moved to
another display.
The same shade window element expanded will be re-expanded after the
shade move is done (either QS or notifications)

The code to do this and wait for the expansion to finish has been
extracted to ShadeExpansionStateInteractor.

As a side effect, this prevents flickers when moving the shade, and
makes the motion much more pleasant.

Bug: 362719719
Test: ShadeDisplaysInteractorTest, ShadeExpandedStateInteractorTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Icacb547aabe613c2b4a2a5d9b96613b300daa87d

Former-commit-id: 362738e68394545cd0fed701a9e808afe202e015