Improve accessibility for notification shelf on lock screen
We added the "Expand" action before to make it clear that the user can
expand the Shelf. But the action didn't have any affect at the time of
addition. This change links the action to the clicking action, which
expands the notification shade. Note that both click and expand actions
have the same effect, which is as intended.

Fix: 378385684
Flag: EXEMPT bugfix
Test: atest Manual, disable "Auto-select" for the Switch Access and
enable Switch Access, select the NotificationShelf on lockscreen. The
action "Expand" should expand the notification shade.

Change-Id: I5b020dbe7c4d79b24b869eb6209b4a59852ba323

Former-commit-id: 6ce4a30b1cd643ee12194b1716e2abfea6f695df