Add new invalid state to print spooler
When a PDF file is invalid (corrupt PDF file, password-protected, etc)
and can't be rendered, switch to an invalid state.  No further updates
happen when the file is known to be invalid and the spooler is in an
invalid state.  This prevents a crash where an update was trying to be
performed when in the finished state.

Additionally, don't allow the user to attempt to restart a print job
with an invalid file since it will simply fail again.

Bug: 389956791
Flag: EXEMPT bug fix
Test: Manual test on akita and brya with password-protected file
Change-Id: I8217463ba3d43a7b898019bdc84443cc43f9d042

Former-commit-id: 97540c5405df864771163386a1e651a70bbcfe6f