PointerIcon: Ensure bitmap size and hotspot are scaled consistently
The density-dependent size of the bitmap is already casted to an int
before we scale it by pointerScale. However, the hotspot did not follow
the same scaling logic, since it was not casted to int before applying
pointerScale. In this CL, we make both the scaling logics consistent.

Bug: 385663202
Change-Id: I375008e5603a9e50e4e7132c8faffda3a1cdd765
Test: manual
Flag: EXEMPT bug fix

Former-commit-id: 626edb8f8c5450fafaee96aaf702cb9e01c55ab8