DisplayTopology: rearrange uses raw overlap values
Fixe a problem in the settings app display topology pane apparent when a
large gap is introduced between two displays in a 3+ display layout. A
large gap results in a very negative xOverlap or yOverlap value. This
should cause X (negative xOverlap) or Y (negative yOverlap) position of
the display to change to make the clamp. Unfortunately the opposite was
happening in extreme cases because we used the absolute x/yOverlap
value. We should use the raw value instead.

Flag: com.android.settings.flags.display_topology_pane_in_display_list
Bug: b/352650922
Test: atest DisplayTopologyTest
Change-Id: Ic0dc32ec10bc76d289c25c24ee02ff600b3afe7c

Former-commit-id: f60636027b0136aa23c9b2208c3e349db8557e91