[Notif redesign] Apply top line margin only to non-headerless notifs
Also move this to applyStandardTemplate instead of
applyStandardTemplateWithActions, since it doesn't actually have
anything to do with actions.

My original assumption was that applyStandardTemplateWithActions is
used for expanded views, but that's wrong on two fronts:
1) it's also used for (collapsed) HUNs and
2) it's NOT used for expanded old media notifs.

As such, instead of doing it in applyStandardTemplateWithActions and
checking for HUNs, we should do it in applyStandardTemplate (which is
called in applyStandardTemplateWithActions too) and make sure we only do
it for non-headerless views. Headerless views use a linear layout to
align content under the top line, so they don't need this special
margin.

This fixes alignment issues in BigPictureStyle and the old
MessagingStyle, as well as the old MediaStyle. HUNs remain the same.

Bug: 378660052
Test: tested manually, screenshot tests to come later
Flag: android.app.notifications_redesign_templates
Change-Id: I274829a12d99ae7ebe4a885c1f9c6828cc25d7ec

Former-commit-id: 3a52013456bdfef66b94ca7f96ae11cb2af8106a