[mobile] Represent invalid subIds as null
MobileConnectionsRepository.defaultDataSubId was allowing
INVALID_SUBSCRIPTION_ID to be propagated downstream, which is kind of
wrong. The only reason we would represent that value was because reading
the int from the broadcast intent wants to have a default value.

This change turns INVALID_SUBSCRIPTION_ID to null downstream, and adds
the necessary changes to support it elsewhere.

Test: StatusBarOperatorNameViewModelTest
Test: MobileConnectionsRepositoryTest
Test: all tets statusbar/pipeline/mobile
Bug: 388834614
Flag: EXEMPT bugfix
Change-Id: I7898aa304b4f2f6b58a88567ed5d42abd65a36f3

Former-commit-id: af694e918c02d9a644fcdc4836e9297da7e4741a