[DnD] Properly cleanup DragDropController in DragDropControllerTests
When deferClose flag was updated to false after the test, it doesn't
fully clean up in some cases.

The idea to set `deferClose=false` is to let `tearDown` clean this up if
`mTarget.dragDropActiveLocked() == true`. However, the definition of
active is not just the presence of dragState. If dragState is present,
but it's closing, it's no longer "active", and hence `closeDragState` is
not executed.

Modify deferClose=false to immediately close dragState.

Bug: 365512241
Test: atest WmTests:DragDropControllerTests
Flag: com.android.window.flags.enable_connected_displays_dnd
Change-Id: Idcbeda929e72a9759a2e3afc103e25a64894ebf5

Former-commit-id: ecb139844d82a1bf4e0a7e73b6dff1cd9a493f5c