Expand correct portion of the shade after moving it to a different display
This fixes the case when one shade element is visible (e.g. QS or notifiactions), but the user swipes down another one in the external display.

Before, the same that was being collapsed was expanded. After, the correct one (based on the status bar touch) is expanded

Bug: 362719719
Bug: 389017437
Test: StatusBarTouchShadeDisplayPolicyTest, ShadeExpandedStateInteractorTest, PhoneStatusBarViewControllerTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Ia7b3ab680a71e16bbe3dfe0d35773195342fa5b0

Former-commit-id: cc32b4c14350f6825ecf150b19466c5da0492c98