Use pending transaction to update client surface on animator
The animation should handle the visibility on container level surface
rather than the client surface.

Especially when the caller is from WindowAnimator which will apply
the pending transaction, the client surface visibility should be
updated in time. Otherwise if a sync state happens before applying
the animator updates the surface, the visibility change will be
deferred until the sync state is finished.

Bug: 387946800
Flag: EXEMPT bugfix
Test: A launching activity with a starting window launches
      another activity in the same task when it is just
      completed drawing. The window of first activity should
      be visible during removing the starting window and
      starting the second transition.
Change-Id: I61694c6d5f94a76e47e9783864f79f370a672a10



Former-commit-id: 33a9c078c6c19bc96a4eed947dffdb60400dbca0