Fix wallpaper dim race condition
The updateListener from updateWallpaperDimming is always run on the UI
thread. The other methods of the engine (here detach()) may run on
another thread if onProvideEngineLooper is overriden, which is the case
for ImageWallpaper.

This can cause a crash if the detach() is called while the wallpaper dim
animation is running. To avoid this, we add a small lock for both the
surface release operations and the wallpaper dim update.

Flag: EXEMPT bugfix
Bug: 386999490
Test: presubmit
Change-Id: I01d014996b0a2df65e17867d1629a211f37f5322

Former-commit-id: e2777eaf0d199613c56424a2a9573829b4eb9362