Fix bouncer dismiss race condition
The bouncer startDisappear flow contains a runnable which must
be consumed in order to tell KeyguardViewMediator to begin hiding
the keyguard. This is supposed to happen after the bouncer disappear
animation plays. However, in rare cases, the runnable gets conflated
with a null value that is intended to clear it.

Use MutableSharedFlow instead of conflating with MutableStateFlow

Test: atest KeyguardBouncerRepositoryTest
Fixes: 383318506
Flag: EXEMPT bugfix
Change-Id: I7c697fdc5c0bbc5b418065f8b65d48aca68b1953

Former-commit-id: 5699d7c2bf5dc0e3d0778801b11685344cfcadef