Remove wtf log for PendingIntent only creator allows BAL
The log was only for Android V to detect any undesired behavior
change. It should now be removed in W. Also removed the flag to
guard the feature in V. Last, revert the change from mAvoidMoveToFront
to mCanMoveToFrontCode that keeps track of reasons to avoid move
to front so that we can log PendingIntent only creator allows.

Bug: 336838334
Bug: 316135632
Test: manual test.
Flag: EXEMPT bug fix
Change-Id: Ib5ce492c19fcca3c8cd73ef2e9608dcab00e95de

Former-commit-id: ac6c0a5c880b06c4e9417ef3cd4c92a5383a820d