Fix issues when shade window is moved very fast between displays
If the shade window was moved too fast between displays the "wait for expansion to finish" was failing with a timeout, making the ShadeDisplayRepository information about where the shade was supposed to be wrong.

In this new version even if we hit the shade expansion timeout we ignore the error (as it can happen also because the user is half expanded manually)

Also cancellation has a better error message that also shows the display id move that failed.

Bug: 362719719
Test: Move the shade between displays quickly and verify the expansion on a new display always triggers the shade to move
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I5b036e8747a523a2ce404dceb175bb42d874f62e

Former-commit-id: 5590a3361707e7600cf05f9a6950800631d7e0e9