Intercept Shade accessibility focus events during launch animations.
This ensures that we don't have stray TalkBack events when a launching
view loses focus after the launch transition is triggered.

Any solution to this problem requires us to rely on the knowledge of
whether or not the launch animation is running. This seems like the
most comprehensive approach for the whole Shade, while being as surgical
as possible and removing the performance implications of marking the
subtree unimportant for accessibility altogether.

Bug: 383688212
Bug: 379222226
Flag: com.android.systemui.shade_launch_accessibility
Test: atest NotificationShadeViewControllerTest NotificationShadeViewTest
Change-Id: I5a65e0107c245e7e8830769f2c515e41578c70bf

Former-commit-id: 8bf1f14005c241c5707f38c4ba12aa40364e6197