Fix bubble bar expanded view content bottom
I previously thought that the content position needs to be adjusted for
insets, but it turns out that the ime position was off. Now that ag/31234392 is submitted and the ime position is correct, we need to fix
the calculation on the bubbles side.

Flag: com.android.wm.shell.enable_bubble_bar
Bug: 377329425
Test: atest BubbleBarExpandedViewTest
Test: manual
       - have a bubble in the bar
       - expand it and request IME
       - verify expanded view is clipped correctly

Change-Id: I8adfcb8e1a8b028180c087448a37a43217a9edf2

Former-commit-id: 638dd788d239f161f4b55ec7c8af193c76fe5362