Show duration dialog when pressing new QS toggle
When turning on DND, if the user has the setting enabled to "Ask every
time", we should show the duration dialog when the new quick toggle in
the Modes tile is pressed. This matches the behavior of the old DND
tile.

Also added logging, and moved the code that creates the dialog to the
delegate to reduce code duplication.

Fix: 388261320
Flag: android.app.modes_ui
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Test: turn on "Ask every time" and check that dialog works correctly
Change-Id: I39644c7748974f56ed87cbcc612854c65b471b1c

Former-commit-id: f6ed4d0644530d065fac8d23d3144e7737456681