Remove obsolete increased heights
These "use increased height" parameters are only true for the old
MessagingStyle or notifications coming from the default messaging
app with category message (see
NotificationMessagingUtil#isImportantMessaging), but it's actually
only used for BigTextStyle. It's an old visual indicator of
"importance" (by showing the notification taller than the rest) that
we don't want to use anymore.

In practice, it would only show a difference for notifications that
fit all the following criteria:
1) uses BigTextStyle
2) has importance LOW or higher (no MIN)
3) has category MESSAGE
4) is coming from the primary SMS app

This is unlikely in practice, as all the main SMS apps use
MessagingStyle, plus we don't actually want to display these
notifications differently now.

Bug: 378660052
Test: builds
Flag: EXEMPT remove mostly-obsolete code
Change-Id: Iab38c1920edf0855aad9c2a7561a31b78193282b

Former-commit-id: ecdcc734116948a518f0567903017cf1925d9578