Transfer Bitmap.mId whenever we send an ashmem bitmap over the wire
Only do this when we do a zero copy transfer of an ashmem region.
Otherwise a value of zero is sent and mId is generated as usual on
the receiving end.

We can look at this in heap dumps and compare identical bitmaps to see
whether they are shared with another process, or are a process-local copy.

Bug: 386212855
Test: boot phone and take heap dumps
Flag: NONE no public API change
Change-Id: Ic791f256884d18f07a03d2f5c7d4e694855e6527

Former-commit-id: 734d655a61aa502eb67acd56eb53b4a02c5fa30d