Correct SQLiteOpenHelper lock creation
This changes SQLiteOpenHelper to synchronize on the raw name of the
database instead of on the canonical file path.

It is possible for an app to create an open helper with two different
db paths that map to the same file on disk; in that case
SQLiteOpenHelper remains thread-unsafe.

Flag: android.database.sqlite.concurrent_open_helper
Bug: 335904370
Bug: 390061581
Test: atest
 * CtsDatabaseTestCases
 * FrameworksCoreTests:android.database
 * CellBroadcastReceiverMTS
Change-Id: I084942d73bdba0adf5a7634b8378e9b397fcc5fa

Former-commit-id: 5253dcb17332c6fcb94e49a53f0dc86c2589d204