Caption/DesktopWindowDecor: don't show window decor for private display.
Shell/SystemUI has no access to private displays, so if a task is
starting on it, even if we get a transition signal to Shell, we should
not do anything such as adding window decoration to it since Shell has
no access to its attributes, such as its context.

This was not a problem since we used to always use the default
DEFAULT_DISPLAY Context.

Bug: 389901828
Test: atest
com.android.test.input.UinputRecordingIntegrationTests#testEvemuRecording
Test: atest
android.server.wm.multidisplay.MultiDisplayKeyguardTests#testDismissKeyguardActivity_secondaryDisplay
Flag: EXEMPT bugfix

Change-Id: I9194d6c89547f18416b4a50e7f3b659c16bef6f8

Former-commit-id: ba537f04e6b3c23b34dba38a469f3f6bb148911b