Rename navigation icon hints to NavbarFlags
The navigation icon hints flags were originally added in [1], with only
the BACK_ALT flag remaining after [2]. Later, the IME visibility and IME
Switcher button visiblity were added for convenience next to this flag,
maintaining the same prefix. However, it is not clear what they hint to,
and they are not related to navigation icons (or even buttons) in
general, but rather to general navigation bar state. While this is
currently only updated from IME state, more flags could be added in the
future, so the IntDef should be general enough to allow this.

  [1]: Ie64369e3df3178178f35b56763ea96d380b6ddd4
  [2]: Ifd1f8c9f400d90542f0ca858b9a4deacabbd518a

Flag: EXEMPT refactor
Bug: 366129400
Test: n/a
Change-Id: I6ec807230f7beb9a9001ec7da08564c7cf1a59ec

Former-commit-id: 4c3eb11709aa755f9235092c8101561238a49ee9