Use cache for volume getter
Using the IpcDataCache to reduce the number of binder calls between
AudioManager and AudioService. From the b/383667500 we see that there
are a lot of calls that can be optimized. We need to investigate whether
the IpcDataCache mechanism is more efficient in this case since the
volumes might get invalidated very often depending on the routing and
mute state.

Flag: android.media.audio.cache_get_stream_volume
Test: atest AudioManagerTest
Bug: 383667500
Change-Id: I8a07d0f728149e2a9bfcaeee81fff60b17ceedb5

Former-commit-id: 4f416c50a7cb590336be31180bf8e3deee2afb4e