Cancel shade window move if there is a new request
This cancels an ongoing shade window request if a new one was issued. It means that instead of waiting for the shade to expand/collapse before moving the window one more time, everything is cancelled and the new request is worked on.

From the user point of view, this results in a much better experience, as in case of shade expansion in a display without the shade, they don't have to wait for the old move + expansion to finish.

Bug: 362719719
Bug: 390371121
Test: ShadeDisplaysInteractorTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: If672187afc6ff5227f4562054c8934e08c747571

Former-commit-id: e0423e86c3a558ee61d5a002ba26675cb9233ea8