Prevent multiple bouncer inflations
Since view inflation happens in an async inflater, it is possible
for multiple copies of the same bouncer view to be inflated.

Check that an inflation has been queued, and if so, add the callback
to a listen that will be invoked when inflation has been completed.

Fixes: 390063829
Test: atest KeyguardSecurityViewFlipperTest KeyguardSecurityContainerControllerTest
Flag: EXEMPT bugfix
Change-Id: I65a012c9b74daf52ee1d4d9643c08a0f08b6d74d

Former-commit-id: 83ee1661598c4b8e44054c3553ec84c152f3b45b