Fix caption bar inset sync issue
When insets are changed (system bar hidden), call to onTaskInfoChanged
happens before DisplayController has the updated value for insets. This
makes the padding in CarWindowDecoration out of sync with the state of
insets. To fix this, listen to inset changes and relayout the decoration
of all visible running tasks.

Bug: 391350956
Test: manual
Flag: com.android.systemui.car.display_compatibility_caption_bar
Change-Id: I3c9eedc252d18a3c5653504941639514a2fb64e2

Former-commit-id: 7280cfcf05f374a6f7337064a65ce4d191d05ad9