Raise the threshold OomAdjuster will drop a service with UI to cached
When a client binds to a proc that has shown UI, OomAdjuster will
aggressively try to drop the service to the cached state to try to
recover memory from the UI process faster. The threshold is based on the
client oom score, but the current threshold is too agressive at
PERCEPTIBLE_APP_ADJ. Raise the threshold to SERVICE_ADJ, the highest
score that can reasonably be performing timely work.

Flag: com.android.server.am.raise_bound_ui_service_threshold
Fixes: 391691057
Test: atest MockingOomAdjuster
Change-Id: Ia17ad5798ef254c9cc1527d85838b833053faf1f

Former-commit-id: e3289ae1d4539772940e9d619e1dea4eff4a8433