[sb] Add new compose battery to HomeStatusBar
When the flag is enabled, this creates and adds a `ComposeView` that
holds a `UnifiedBattery` in the status bar

It also turns off the old BatteryMeterViewController, which gets
injected around. Once the flag rolls out, we can delete all of that
nonsense.

Test: manual
Bug: 391605373
Flag: com.android.settingslib.flags.new_status_bar_icons
Change-Id: I771fe3a80d2b46ac9cdf921d053daf115aeb6532

Former-commit-id: 9a984b49b627ed3eb9c7921da35fac592f90fa9a