Fix: Prevent Voice Access from highlighting non-clickable area
This commit fixes an issue where Voice Access was incorrectly highlighting a non-clickable element as clickable.

The issue occurred because the non-clickable area had a no-op click listener attached to it, which was being detected by Voice Access as a clickable element.

To resolve this, mark the view as non important for A11Y.

Change-Id: I501bfaaa00c03905b6cb075d30627035193be3d0
Fixes: 380155292
Test: manual
Flag: NONE trivial fix

Former-commit-id: 40e12ae8902500c7f759492163598c0bafc552fb