Update workchain when wakelock uids updated
There are currently 2 ways to update the worksource with the caller uids
for appropriate attribution - Workchain and uids array in the WorkSource
object. The array approach has been deprecated, and is a hidden utility.
Workchain is the recommended way to do the attribution. A special hack
was created for Audio wakelocks to create wakelocks in async, and then
update the corresponding uids using updateWakelockUids API. This API
would then update the uids array inside the Worksource, and hence the
changing uids never get logged in the form of atoms.

Bug: 331304805
Flag: com.android.server.power.feature.flags.wakelock_attribution_via_workchain
Test: atest PowerManagerServiceTest

Change-Id: I4d1b8559c4293c35be5ba8d30957b5b48311c1ab

Former-commit-id: ee0d554cd57ad1fefe23a945dc7852fde78b7adf