Default dream trigger to postured
The current default is that dreams are enabled by default but only
enabled "while docked". On devices that support posturing, we will
switch the default to be "while postured" and disable the main dream
toggle to maintain existing behavior of the dream not being enabled.

If the user has switched their setting to "while charging", no changes
will be made.

Bug: 383208131
Test: flashed change with wiping device and without wiping device, to
ensure the upgrade works correctly in both scenarios.
Flag: EXEMPT settings database upgrade

Change-Id: Idba32d9a83b163ae0a3d156c12849c584290c63e

Former-commit-id: 68fcf43c55fbae84838902d14984f4961ae9c42c