Made spacing between keyboard settings text and icon constant.
In some locales where the translation of keyboard settings was too long,
the space between keyboard setting and the icon was virtually
non-existent. And If the translation was slightly longer, the icon
disappeared entirely, pushed out of the layout.
This change corrects this by making the text have a weight of 1 while
leaving the space and icon unweighted. This implies that the space and
icon will be measured first and laid out, then the text will take up the
remaining space.

Flag: com.android.systemui.keyboard_shortcut_helper_shortcut_customizer
Test: Manual
Fix: 386754676
Change-Id: Iaec07e7f01835c4eff33f13497944bcf6306078e

Former-commit-id: 04a85b1091ddfe3b267e01fcaa73e898da314ddf