Make long-lived transition animations asynchronous.
This is because the necessary Expandable for a given transition might
not be composed when the animation is triggered. It is up to the
component owning the transition to make sure that a composition
happens that includes the Expandable, so it can then be used and the
animation can start.

Bug: 323863002
Flag: com.android.systemui.shared.return_animation_framework_library
Flag: com.android.systemui.shared.return_animation_framework_long_lived
Test: atest ActivityTransitionAnimatorTest ActivityStarterImplTest LegacyActivityStarterInternalImplTest
Change-Id: Ie0d9c47b5c4cb89f0ca3194bf325e4da6fcb3819

Former-commit-id: 1ca9e829028c430d281abdc326c4113d7f0b7b8f