[flexiglass] Adds support for trust agents
When the environment is trusted, the device shouldn't become locked when
it goes to sleep or starts dreaming. At least not until the environment
becomes untrusted again or a more explicit request to lock happens (for
example, a lock down).

Fix: 392181543
Test: unit tests added
Test: manually verified, using bluetooth headphones that have been set
up as a trusted device in the "Extend unlock" feature in settings, that (1) the device
remains unlocked while the headphones are present for up to the
timeout/limit (4 hours normally but locally changed to 60 seconds by
editing TRUST_TIMEOUT_IN_MILLIS in TrustManagerService). Also verified
that bringing up the power menu makes the environment untrusted and
that disconnecting the trusted device also does.
Flag: com.android.systemui.scene_container

Change-Id: I81a5e0fa2ac3b5b6c858274d4af478efb8daaeb0

Former-commit-id: 42f1f7626f192a9ac41f73517390491a13d668a9