Use the right overlay when using an Expandable's transition controller.
I noticed in my prototyping that sometimes `composeViewRoot.rootView`
here is different from the one provided as the `transitionContainer`.
This causes the drawable and the ghost view to be drawn in two
different overlays, which can result in the view never being covered
by the drawable. Reusing the already set `transitionContainer`
guarantees that the two are in the same overlay and that their
respective order is correct.

Bug: 323863002
Flag: EXEMPT bugfix
Test: I manually tested the current Compose launches (that I'm aware of) as well as the new call chip integration in my prototype.
Change-Id: I8e6f417e5a7a7a1e1fec37b1f61579a9335db732

Former-commit-id: 599ccdb975894e78136db0a9ee11d78de6a9f8e3