Change BubbleTaskViewHelper to implement the listener
The helper was really the listener plus some extra stuff
around state that's relevant to the listener.

This CL makes it so that the listener is the top level class
making it easier to test.

Adds a test for BubbleTaskViewHelper

Currently this is only used by bubble bar expanded view.
Will migrate bubble expanded view to this in a separate CL.

Will rename in a separate CL

Flag: com.android.wm.shell.enable_bubble_bar
Bug: 385361517 (proper tests for the fix for this bug)
Bug: 272102927
Test: atest BubbleTaskViewHelperTest
Change-Id: I92de8b1838e3ef2bc18d0b55723a216cc876e27a



Former-commit-id: e5a5e58625e793bab5da9d3a407d02fcca308010