[ECM] Change default mode of ACCESS_RESTRICTED_SETTINGS
This sets the default mode of the ACCESS_RESTRICTED_SETTINGS op to
MODE_DEFAULT, and removes the now-unnecessary call to set the mode
from InstallPackageHelper.

Also, since app ops essentially stores all default modes values as
null, we'll have to do an upgrade step. For example, the previous
default mode (for this op) was MODE_ALLOWED, which was stored in app
ops as null. But, since we're redefining the default mode from
MODE_ALLOWED to MODE_DEFAULT, then we'll have an issue if we take no
further action, because all null values will suddenly mean MODE_DEFAULT
instead of MODE_ALLOWED, which is not desirable, because while we do
want to henceforth change the default mode for all newly installed apps,
we also want already-installed app op mode values to retain their
meaning. The solution is to perform a one-time upgrade of all null
values and resetting those values explicitly to MODE_ALLOWED.

Bug: 388960315
Test: manual
Test: EnhancedConfirmationManagerTest
Test: EnhancedConfirmationInCallTest
Test: CtsPackageInstallerCUJInstallationTestCases:android.packageinstaller.criticaluserjourney.cts.InstallationViaSessionTest
Relnote: N/A
Flag: EXEMPT bugfix
Change-Id: I53fbb8b23a404f790f3b0f71bdc2dbdbafac392d

Former-commit-id: 44c4665074fbb9327b25b298209f99488a5d0910