Make sure Bitmap for PiP overlay being recycled
- In this change

Add a last resort as to recycle the bitmap being used in
PipAppIconOverlay. Currently, we rely on the creator of
PipAppIconOverlay to call detach and that practice is not strictedly
followed in both Launcher and PiP1 implementation.

- What's next

We will evaluate if a proper fix is worthwhile in PiP1 and meanwhile,
make sure PiP2 implementation does the recycle properly.

Flag: EXEMPT bugfix
Bug: 393584443
Test: Enter PiP with app icon overlay, make sure bitmap is recycled
Change-Id: Ie8981bd1e52f53bfe010aa5c9bd3ff6f69f30c33

Former-commit-id: ce298705433b7fccbd084c4fbb3aadc911b6a1eb