Introduce bubble types & clean up some naming
When creating a bubble, a type is set internally.
It can be chat, note, app, or shortcut.

The type is based on the constructor content.

The constructors are a bit messy and I'll be cleaning
them up in a future change.

This CL adds methods to check the type and updates
some of the other related naming to be more consistent.

In some places isAppBubble was really checking both
app & notes, this is now made more explicit by checking
for both.

Adds tests for the types.

Fixes up a test in BubblesTests where you could make a
non-chat notification-based bubble, which doesn't exist.

Flag: EXEMPT minor refactor
Test: atest BubbleTest
Bug: 390496034
Change-Id: I3b85ea4d607bc81cd5b45632e3b07d372fe30f54



Former-commit-id: c0d7e90025c9680060de57dfe2e8383248c6a7a4