Pipe depth signal to keyguard for spatial model zoom out
Lockscreen should be pushed back when the shade is pulled down. The pushback level is 5%, the same as the homescreen.

For the most part of the spatial model work, the implementation is done in WMShell. It scales down the entire display area while filtering out a few specific windows (e.g. nav bar, status bar, shade). Since the lockscreen and the shade share the same window, pushing back the lockscreen has to be done at the view level (we are scaling the root view of the lockscreen here).

Bug: 391901376
Test: video in the bug
Flag: com.android.systemui.spatial_model_app_pushback
Change-Id: I6aad1bf074a43949644f03ef738b392a7865b504

Former-commit-id: 16bb8044eda5ee42154a586dba3ad795afc6bc49