Wait for IME hide in verifyFullscreenMode
The testFullScreenMode has one scenario for each of natural orientaiton,
left orientation and right orientation. Each of these triggers an IME
show, waits for it to complete, and then presses the back key to hide
the IME, however without waiting for this to take effect.

We use the same countDownLatch to verify either IME show, IME hide or
IME rotation change. When going from left to right orientation, we don't
expect a rotation change. However, the hide request from the left
orientation is still ongoing when moving to right orientation, and will
trigger the countDownLatch, failing the test.

This fixes the test by waiting for the IME visibility to change at the
end of each scenario.

Additionally, this removes the dependency of finding and clicking on the
EditText, and pressing the back key, by using the insets APIs to show
and hide the IME, reducing the test time. To make this work we have to
get the newly created Activity after the configuration change took
effect.

Flag: EXEMPT testfix
Bug: 393025053
Test: atest InputMethodService#testFullScreenMode
Change-Id: I3bfb84e37e198ff89fd9e7e36a2e2236c5571b36

Former-commit-id: 37efb549ed72c8f1d38db77de72ac4484234e49e