Estimate vibration effect size using available parcel data
While we want to avoid reading in vibration effects that are too large (which can cause the XML serialization to fail), serializing directly in the constructor to check the length causes too much heap churn. This change uses the remaining data available in the parcel to estimate the size of the vibration effect, after moving it to the end of the parcel.

Flag: android.app.notif_channel_estimate_effect_size
Bug: 391908451
Test: heapdump, NotificationChannelTest
Change-Id: Ic58addd01721c8a0f358deb8629302a3d57e7673

Former-commit-id: 6875f99a68d797ff70a4b224638c67283dfe7ff3