Disable nested drag only in disableSwipesWhenScrolling()
This CL is changing the implementation of disableSwipesWhenScrolling()
so that only the nested scroll behavior of Modifier.nestedDraggable() is
disabled when scrolls have been consumed by a nested scrollable, instead
of disabling the entire PointerInputModifierNode and removing it from
composition.

This change is necessary because Compose does not really support the
removal of a PointerInputModifierNode during a gesture (see b/389906055
for details).

Bug: 389906055
Test: atest NestedDraggableTest
Test: atest ContentTest
Flag: com.android.systemui.scene_container
Change-Id: I6b0bca8f6915c8f7acf67ff81241ae36daf30dd2

Former-commit-id: 880ef4fa59535266bcf8db0ec23717eb857b2998