[pm] move PackageInstallerSession.extractNativeLibraries to IO thread
PackageInstallerSession.write() holds mLock and is on the IO thread.
Prior to this CL, PackageInstallerSession.extractNativeLibraries() also
holds mLock but was on the mInstallThread. This causes a lock contention
where the write() can ANR waiting for mLock held by
extractNativeLibraries(). Move the latter to the IO thread to avoid such
lock contentions.

FLAG: EXEMPT refactor
BUG: 393698598
Test: manual

Change-Id: I99418b8983a63751e6d485efae947a49a6a88db5

Former-commit-id: bac99b5f3bfbf6cdb475d072125ac88327f66806